requirements.txt
slackclient==2.1.0